requirements.txt
streamlit
requests
pytesseract
Pillow
openai-whisper
matplotlib
pydub
langchain
ollama
langchain-community
transformers
torch
